Move tests to a common abstract test class
The tests from async executor, sync executor and Default resolver are
basically the same, changing just the executor used for the tests. This
commit moves all the tests to an abstract base class to avoid code
duplication. The DefaultSPFResolverTest will run the same test as sync
executor test class, but using the DefaultSPF constructor.